apps/backend/requirements.txt
Django>=4.2,<5.0
psycopg2>=2.9.5
django-cors-headers>=4.0.0
djangorestframework>=3.14.0
python-decouple>=3.6
requests>=2.28.0
pytz>=2023.3
PyJWT>=2.6.0 